rename: molecules/banner/Main -> molecules/banner/SearchBanner
