+++ molecules/banner/SearchBanner.module.scss
