+++ molecules/banner/index.ts
@@ -1,1 +1,1 @@
-export { default as Banner } from './Banner';
+export { default as SearchBanner } from './SearchBanner';
