+++ pages/index.tsx
@@ -1,13 +1,13 @@
 import classNames from 'classnames';
 import { Container } from '~/atoms/layout';
 import { MostViewSection, FeaturedSection } from '~/molecules/preview';
-import { Banner } from '~/molecules/banner';
+import { SearchBanner } from '~/molecules/banner';
 import styles from './index.module.scss';
 
 export default function MainPage() {
   return (
     <>
-      <Banner />
+      <SearchBanner />
       <Container className={styles.contents}>
         <div className={styles.mostViewWrapper}>
           <div className={classNames(styles.column, styles.left)}>
